+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### v1.0.0
-* Initial release
+* Initial release.
 
 ### v1.0.1
 * Edited documentation a bit.
@@ -12,13 +12,13 @@
 * Added add_chat_message (from ShareItem).
 
 ### v1.0.4
-* find_item works with "namespace-identifier" strings now
+* find_item works with "namespace-identifier" strings now.
 * Added: get_host_player
 * Added: is_lobby_host
 * Added: table_merge
 
 ### v1.0.5
-* get_client_player works in Trials mode now
+* get_client_player works in Trials mode now.
 
 ### v1.0.6
 * Added: table_to_string
@@ -31,7 +31,15 @@
 
 ### v1.0.7
 * Added: is_singleplayer_or_host
-* net_send can now exclude a player
+* net_send can now exclude a player.
 
 ### v1.0.8
 * Added: net_clear
+
+### v1.0.9
+* Added: is_lobby_client
+* Added: spawn_crate
+* string_to_table now reads nil correctly.
+* find_item and get_all_items now return new copies of the item data tables.
+    * This is to prevent direct manipulation of the main one, which is shared by all mods.
+* Net functionality now intercepts messages instead of scanning chat and deleting.